fix: [minor] Avoid unnecessary typecasting. (#4828)
We don't need to typecast identifiers from
their base to type to same type again. This
is not a bug and compiler is fine to skip
it but it is better to avoid if not needed.

+++ cmd/erasure-readfile.go
@@ -75,8 +75,8 @@ func (s ErasureStorage) ReadFile(writer io.Writer, volume, path string, offset, 
 			return f, err
 		}
 		startOffset = 0
-		f.Size += int64(n)
-		length -= int64(n)
+		f.Size += n
+		length -= n
 	}
 
 	f.Algorithm = algorithm
